pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sable-platform"
version = "0.1.0"
description = "Suite-level platform backbone for the Sable tool stack"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
sable-platform = "sable_platform.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["sable_platform*"]

[tool.setuptools.package-data]
sable_platform = ["db/migrations/*.sql"]

[tool.pytest.ini_options]
testpaths = ["tests"]
